Add mention of `fromIntegral` function.
Per #28 , added some wording for the `fromIntegral` function.

Removed the 'aside' for `ghcid` as this was causing more problems
than it was solving. Too many install / usage dramas.

Added a snippet for loading a repl with test libraries and packages
loaded. Because I had completely forgotten how to do that.